Writing 파일 폴더링 / Extension 파일 세분화 하여 폴더링
